requirements.txt
potassium>=0.4
pyannote.audio>=3